site/requirements.txt
alabaster==0.7.10
anaconda-client==1.6.3
anaconda-navigator==1.6.2
anaconda-project==0.6.0
asn1crypto==0.22.0
astroid==1.4.9
astropy==1.3.2
Babel==2.4.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.0
bitarray==0.8.1
blaze==0.10.1
bleach==1.5.0
bokeh==0.12.5
boto==2.46.1
Bottleneck==1.2.1
branca==0.2.0
Cerberus==0.9.2
certifi==2016.2.28
cffi==1.10.0
chardet==3.0.3
click==6.7
cloudpickle==0.2.2
clyent==1.2.2
colorama==0.3.9
contextlib2==0.5.5
cryptography==1.8.1
cycler==0.10.0
Cython==0.25.2
cytoolz==0.8.2
dask==0.14.3
datashape==0.5.4
decorator==4.0.11
distributed==1.16.3
Django==1.9.4
django-filter==0.15.2
djangorestframework==3.4.7
docutils==0.13.1
entrypoints==0.2.2
et-xmlfile==1.0.1
Eve==0.7.8
Events==0.2.2
fastcache==1.0.2
Flask==0.12
Flask-Cors==3.0.2
Flask-PyMongo==0.5.1
folium==0.5.0
future==0.16.0
geocoder==1.33.0
geojson==2.3.0
geojson-utils==0.0.2
geopy==1.11.0
gevent==1.2.1
gmaps==0.7.0
googlemaps==2.5.1
greenlet==0.4.12
h5py==2.7.0
heamy==0.0.7
HeapDict==1.0.0
html5lib==0.9999999
idna==2.5
imagesize==0.7.1
ipykernel==4.6.1
ipython==5.3.0
ipython-genutils==0.2.0
ipywidgets==7.0.5
isort==4.2.5
itsdangerous==0.24
jdcal==1.3
jedi==0.10.2
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.0.1
jupyter-console==5.1.0
jupyter-core==4.3.0
Keras==2.1.5
lazy-object-proxy==1.2.2
lightgbm==2.0.11
llvmlite==0.18.0
locket==0.2.0
lxml==3.7.3
Markdown==2.6.9
MarkupSafe==0.23
matplotlib==2.0.2
mistune==0.7.4
mpmath==0.19
msgpack-python==0.4.8
multipledispatch==0.4.9
navigator-updater==0.1.0
nbconvert==5.1.1
nbformat==4.3.0
networkx==1.11
nltk==3.2.3
nose==1.3.7
notebook==5.0.0
numba==0.33.0
numexpr==2.6.2
numpy==1.14.2
numpydoc==0.6.0
odo==0.5.0
olefile==0.44
openpyxl==2.4.7
packaging==16.8
pandas==0.20.1
pandas-profiling==1.4.0
pandocfilters==1.4.1
partd==0.3.8
pathlib2==2.2.1
patsy==0.4.1
pep8==1.7.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==5.1.0
plotly==2.0.11
ply==3.10
prompt-toolkit==1.0.14
protobuf==3.4.0
psutil==5.2.2
ptyprocess==0.5.1
py==1.4.33
pycosat==0.6.2
pycparser==2.17
pycrypto==2.6.1
pycurl==7.43.0
pyflakes==1.5.0
Pygments==2.2.0
pylint==1.6.4
pymongo==3.6.1
pyodbc==4.0.16
pyOpenSSL==17.0.0
pyparsing==2.1.4
pytest==3.0.7
python-dateutil==2.6.0
pytz==2017.2
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==16.0.2
QtAwesome==0.4.4
qtconsole==4.3.0
QtPy==1.2.1
ratelim==0.1.6
requests==2.14.2
rope-py3k==0.9.4.post1
scikit-image==0.13.0
scikit-learn==0.18.1
scipy==0.19.0
seaborn==0.7.1
Shapely==1.6.3
simplegeneric==0.8.1
simplejson==3.13.2
singledispatch==3.4.0.3
six==1.11.0
snowballstemmer==1.2.1
sortedcollections==0.5.3
sortedcontainers==1.5.7
Sphinx==1.5.6
spyder==3.1.4
SQLAlchemy==1.1.9
statsmodels==0.8.0
sympy==1.0
tables==3.3.0
tblib==1.3.2
tensorflow-gpu==1.3.0
tensorflow-tensorboard==0.1.6
terminado==0.6
testpath==0.3
toolz==0.8.2
torch==0.3.1
torchvision==0.2.0
tornado==4.5.1
tqdm==4.19.4
traitlets==4.3.2
transliterate==1.10
unicodecsv==0.14.1
vincenty==0.1.4
wcwidth==0.1.7
Werkzeug==0.12.2
widgetsnbextension==3.0.8
wrapt==1.10.10
xgboost==0.6
xlrd==1.0.0
XlsxWriter==0.9.6
xlwt==1.2.0
zict==0.1.2
